requirements.txt
openslide-python==0.5.1
argparse==0.3.7
Pillow==2.3.0
